refactor: refactor auth methods in authfacade class
