+++ backend/src/main/java/net/tamasnovak/utilities/authenticationFacade/AuthenticationFacade.java
@@ -1,8 +1,14 @@
 package net.tamasnovak.utilities.authenticationFacade;
 
 import net.tamasnovak.entities.account.Account;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.User;
 
 public interface AuthenticationFacade {
+  Authentication authenticateUser(String email, String password);
   Account getAuthenticatedAccount();
   String getAuthenticatedAccountRole();
+  User getUserContext();
+  User getUserDetails(Authentication authentication);
+  String transformRolesArrayToString(User userDetails);
 }
