+++ backend/src/main/java/net/tamasnovak/utilities/authenticationFacade/AuthenticationFacadeImpl.java
@@ -2,21 +2,31 @@ package net.tamasnovak.utilities.authenticationFacade;
 
 import net.tamasnovak.entities.account.Account;
 import net.tamasnovak.services.account.account.AccountService;
-import net.tamasnovak.utilities.StringFormatterUtilities;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Component;
 
 @Component
 public class AuthenticationFacadeImpl implements AuthenticationFacade {
+  private final AuthenticationManager authenticationManager;
   private final AccountService accountService;
-  private final StringFormatterUtilities stringFormatter;
 
   @Autowired
-  public AuthenticationFacadeImpl(AccountService accountService, StringFormatterUtilities stringFormatter) {
+  public AuthenticationFacadeImpl(AuthenticationManager authenticationManager, AccountService accountService) {
+    this.authenticationManager = authenticationManager;
     this.accountService = accountService;
-    this.stringFormatter = stringFormatter;
+  }
+
+  @Override
+  public Authentication authenticateUser(String email, String password) {
+    UsernamePasswordAuthenticationToken auth = new UsernamePasswordAuthenticationToken(email, password);
+
+    return authenticationManager.authenticate(auth);
   }
 
   @Override
@@ -26,10 +36,27 @@ public class AuthenticationFacadeImpl implements AuthenticationFacade {
 
   @Override
   public String getAuthenticatedAccountRole() {
-    return stringFormatter.transformRolesArrayToString(getUserContext());
+    return transformRolesArrayToString(getUserContext());
   }
 
-  private User getUserContext() {
+  @Override
+  public User getUserContext() {
     return (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+  }
+
+  @Override
+  public User getUserDetails(Authentication authentication) {
+    return (User) authentication.getPrincipal();
+  }
+
+  @Override
+  public String transformRolesArrayToString(User userDetails) {
+    String role = userDetails.getAuthorities()
+      .stream()
+      .map(GrantedAuthority::getAuthority)
+      .toList()
+      .toString();
+
+    return role.substring(1, role.length() - 1);
   }
 }
